# docs/swagger.yaml
definitions:
  api.DependencyStatsResponse:
    properties:
      by_type:
        additionalProperties:
          type: integer
        type: object
      top_dependencies:
        items:
          $ref: '#/definitions/database.Dependency'
        type: array
      total_dependencies:
        example: 150
        type: integer
    type: object
  api.ErrorResponse:
    properties:
      error:
        example: Invalid request
        type: string
    type: object
  api.ListProjectsResponse:
    properties:
      count:
        example: 5
        type: integer
      projects:
        items:
          $ref: '#/definitions/api.ProjectSummary'
        type: array
    type: object
  api.ListReportsResponse:
    properties:
      count:
        example: 3
        type: integer
      reports:
        items:
          $ref: '#/definitions/database.Report'
        type: array
    type: object
  api.ProjectSummary:
    properties:
      id:
        example: 1
        type: integer
      name:
        example: My Project
        type: string
      repo_url:
        example: https://github.com/owner/repo
        type: string
    type: object
  api.SubmitRequest:
    properties:
      description:
        example: A sample project
        type: string
      github_token:
        example: ghp_xxxxxxxxxxxx
        type: string
      name:
        example: My Project
        type: string
      repo_url:
        example: https://github.com/username/repo
        type: string
    required:
    - repo_url
    type: object
  api.SubmitResponse:
    properties:
      message:
        example: Report generation started
        type: string
      project_id:
        example: 1
        type: integer
      report:
        $ref: '#/definitions/database.Report'
      report_id:
        example: 1
        type: integer
    type: object
  api.UpdateProjectRequest:
    properties:
      description:
        example: Updated description
        type: string
      github_token:
        example: ghp_xxxxxxxxxxxx
        type: string
      name:
        example: Updated Project Name
        type: string
      regenerate:
        example: true
        type: boolean
    type: object
  database.Dependency:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      package_type:
        description: Unique combination of package type, name, and version
        type: string
      repo_url:
        description: Repository information
        type: string
      updated_at:
        type: string
      version:
        type: string
      vuln_count:
        description: Vulnerabilities
        type: integer
    type: object
  database.Project:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      repo_url:
        type: string
      reports:
        items:
          $ref: '#/definitions/database.Report'
        type: array
      updated_at:
        type: string
    type: object
  database.Report:
    properties:
      base_dir:
        type: string
      created_at:
        type: string
      dependencies:
        items:
          $ref: '#/definitions/database.Dependency'
        type: array
      generated_at:
        type: string
      graph_svg:
        description: SVG graph
        type: string
      html_report:
        description: HTML report
        type: string
      id:
        type: integer
      project:
        $ref: '#/definitions/database.Project'
      project_id:
        type: integer
      sbom_data:
        description: JSON or XML data
        type: string
      sbom_format:
        description: SBOM data
        type: string
      total_dependencies:
        description: Stats
        type: integer
      total_vulns:
        type: integer
      updated_at:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/v1/dependencies:
    get:
      description: Returns all unique dependencies across all projects (deduplicated)
      parameters:
      - description: Filter by package type (npm, python, go, maven)
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.Dependency'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: List all dependencies
      tags:
      - dependencies
  /api/v1/dependencies/stats:
    get:
      description: Returns statistics about dependencies across all projects (deduplicated)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.DependencyStatsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get dependency statistics
      tags:
      - dependencies
  /api/v1/projects:
    get:
      description: Returns a minimal list of all submitted projects (ID, name, URL
        only)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ListProjectsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: List all projects
      tags:
      - projects
  /api/v1/projects/{id}:
    get:
      description: Returns details of a specific project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.Project'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get a project by ID
      tags:
      - projects
    put:
      consumes:
      - application/json
      description: Update project details and optionally regenerate SBOM report with
        new GitHub token
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.UpdateProjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SubmitResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Update a project and optionally regenerate report
      tags:
      - projects
  /api/v1/projects/{id}/regenerate:
    post:
      description: Triggers regeneration of SBOM report using stored GitHub token
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SubmitResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Regenerate SBOM report for a project
      tags:
      - projects
  /api/v1/projects/{id}/reports:
    get:
      description: Returns all reports generated for a specific project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ListReportsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: List reports for a project
      tags:
      - reports
  /api/v1/reports/{id}:
    get:
      description: Returns details of a specific report including dependencies
      parameters:
      - description: Report ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.Report'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get a report by ID
      tags:
      - reports
  /api/v1/reports/{id}/graph:
    get:
      description: Returns the SVG dependency graph for a specific report
      parameters:
      - description: Report ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - image/svg+xml
      responses:
        "200":
          description: SVG content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get report dependency graph
      tags:
      - reports
  /api/v1/reports/{id}/html:
    get:
      description: Returns the HTML report for a specific report
      parameters:
      - description: Report ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - text/html
      responses:
        "200":
          description: HTML content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get report HTML
      tags:
      - reports
  /api/v1/submit:
    post:
      consumes:
      - application/json
      description: Analyzes a Git repository and generates an SBOM report
      parameters:
      - description: Repository information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.SubmitRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SubmitResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Submit a repository for SBOM analysis
      tags:
      - repositories
  /health:
    get:
      description: Returns the health status of the API
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Health check endpoint
      tags:
      - system
swagger: "2.0"